mtlab-valid-end-construct-p - Add save-match-data around body.
matlab-up-string-orocomment, and matlab-backward-up-string-or-comment - Return t if we moved.
matlab-move-list-sexp-internal - skip whitespace before moving, and do
special stuff we we start in a string or comment, but are at the beginning/end of it.
matlab-backward-sexp - fix some cases with 'noerror' feature.
matlab-lattr-block-close - Fix to skip over ends in comments, strings and other places
while still counting ends that are valid.
matlab-show-paren-or-block - Fix case when on end to ignore ends that are array indices,
and use 'noerro' intput to matlab-backward-sexp.