matlab.el:
(matlab-font-lock-extend-region):
If we extend, extend for whole lines.
(matlab-find-block-comments):
On fail or success, explicitly move to end
of what we have found so far.
(matlab-ltype-block-comm): Reject block
comment starts in strings or other comments.
Track if we found a comment successfully
during search loops, and don't continue if that
failed, but we had a success on an earlier loop that
was in a string or comment.
Support new input to extend to line bounds.