matlab-topic.el:
Fix to work w/ newer versions of Emacs.
(matlab-shell-help-mode): Add mouse-1 keybinding for clicking on things.
(mtlab-shell-topic-browser-create-contents): Replace toggling
read-only-mode with `inhibit-read-only' feature.
Add bit @ end to force font-lock to activate for this buffer.
In E27 it seems to not colorize.
(matlab-shell-topic-choose): Don't error if not clicking on a topic now
that mouse-1 also clicks in addition to moving point.
(matlab-shell-topic-mouse-highlight-subtopics):
Use `inhibit-read-only' when adding text properties.