toolbox/+emacs/@Emacsserver/EmacsServer.m
Add FollowStack option that can be turned on/off.
When on, get the Stack object, and update the stack.

toolbox/+emacs/@Stack/Stack.m
Trim off all the timer cb parts of the stack to get to
where the user code really is in the case of EmacsServer.
Fix file names on Windows to have Emacs friendly slashes.

toolbox/+emacs/set.m
Add 'followstack' option.
Sets FollowStack prop on the Server.