toolbox/emacsnetshell.m:
Add fetch option to just return the server if there already is one, or empty.
Support calling this function with no argument after it is running, and doing a 'fetch'
instead of an init.

help.m:
Use 'fetch' of emacsnetshell to decide if we need to send cookies to Emacs.
Don't send those cookies with netshell - only if inside an Emacs buffer.