matlab.el:
Speed up checking for valid class block keywords.
(matlab-find-prev-code-line): New.
Looks for a previous line of code and navs to the beginning of it.
(matlab-cursor-on-valid-block-start):
Stop navigating for full syntactic block check, just look at
prev command and verify it belongs to a class and assume it is correct.
(matlab-previous-line-belongs-to-classdef-p):
New.  Verify previous thing is a class based keyword.
(matlab-calculate-indentation-1):
When indenting `end' ask matlab-backward-sexp to not throw errors.