matlab.el:
(matlab-valid-end-construct-p): an 'end' after a . is a field,
and not a valid block end.
(matlab-cursor-on-valid-block-start): any block after a . is a field,
and not a valid block start.

mlint.el:
(mlint-lm-missing-end::mlint-fix-entry): Fix too many close )

tests/blocks.m:
Add struct_stuff method to verify not indenting anything after a .

tests/stringtest.m:
Add an invalid block comment (all on one line) with test markers in it.