(matlab-beginning-of-command): Remove "P" from interactive cmd as this has no inputs anymore.
(matlab-indent-region): Add "r" to interactive cmd so it can get the active region.
(matlab-set-comm-fill-prefix): Add special case for prefix for comments after code.
(matlab-auto-fill): Get the end-comment-column and not just a comment on the line.
(matlab-fill-comment-line): Replace code with call to 'fill-paragraph'.
(matlab-fill-paragraph): For comments, use beginning/end of string-or-comment instead of
command to get real start/end of the whole comment.
Always start from beginning of the comment (not the middle) to compute prefix.
Fill using `fill-comment-paragraph' not more generic `fill-paragraph'.