toolbox/help.m: simplify implementation
M-x matlab-shell will addpath /path/to/Emacs-MATLAB-Mode/toolbox which contains help.m
that shadows the built-in help. The role of help.m is to wrap the results of the built-in
help.m in <EMACSCAP>(bufname) HELP </EMACSCAP> which and matlab-shell then places HELP
in a bufname Emacs buffer.

To get at the built-in help, we were manipulating the MATLAB path, but under some cases
we were failing to run the built-in help.

This commit levarages which('-all','help.m') to locate the built-in help.m and then
verifies we found the built-in help.m.