make: run tests with a clean matlab setup
If the user has ~/matlab/startup.m or MATLABPATH environment variable setup, these could
alter the behavior of MATLAB and cause tests to fail.

Therefore, we add tests/startup.m which is empty and unset the MATLABPATH prior to running the
tests. With this in place, the tests pass even when running with
   echo "quit" > ~/matlab/startup.m
   or bad MATLABPATH

Also updated the DEBUGGING.org to suggest these tips for running with an unaltered MATLAB.